+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/PurchaseCancelServiceImpl.java
@@ -182,7 +182,7 @@ public class PurchaseCancelServiceImpl implements PurchaseCancelService {
                 purchase.getUserId(),
                 tickets,
                 merchandises,
-                purchase.getTotalPrice(),
+                purchase.getTotalPrice().longValue(),
                 purchase.getPurchaseDate(),
                 purchase.getStreet(),
                 purchase.getPostalCode(),
